requirements.txt
pandas==1.5.3
streamlit==1.17.0
plotly==5.13.0
pillow==9.4.0